* Overhaul equip drop table Most changes affect the pre-110 drops to drop all of the Novice/1st-4th Class sets' armor pieces for newbies. I haven't really touched on the other in-between gaps besides filling in gaps (incl. a handful of weapons and matching tops/bottoms for either/both genders), and removing 15/25/35 gear from the drop table (these are "supposed" to be from shops only, but the shop SQL as a whole also needs to be overhauled). * Purge specific pot drops from drop table, adjusted initConsumableDrops' boundaries Evidentially I didn't get the memo from initConsumableDrops under ItemConstants when overhauling general mob drops. In GMS, the pot level ranges are defined as follows: 0-30, 31-60, 61-100, 101-140, and 141+. I don't really want to try refactoring getConsumableMobDrops drastically to try matching those, so instead I've just shifted the current pot distributions and bumped up the upper bound to match those ranges more closely. * Reorganize JobConstants' basic isClass functions and refactored each Explorer base class' isClass functions * Streamline getMainStatForJob and getItemJobByJob, fixed Battle Mage being missing from both functions and Xenon not being in getItemJobByJob Stop reinventing the wheel thx * Fix Corsair's main/secondary stat definitions being flipped around isAdventurerPirate is defined in the STR getMainStatForJob if statement...and the former function also covers Corsair's job ids. Since that return is reached first instead of Corsair's intended DEX return, Corsair's main/secondary stats ended up being backwards. I've replaced that function call with a basic job == 500 to cover 1st job Explorer Pirates instead.

+++ src/main/java/net/swordie/ms/constants/SkillConstants.java
@@ -417,7 +417,7 @@ public class SkillConstants {
             return 2333001;
         if ( JobConstants.isMercedes((short) job) )
             return 2061010;
-        if ( JobConstants.isAngelicBuster(job) )
+        if ( JobConstants.isAngelicBuster((short) job) )
             return 2333001;
         // TODO:
 //        if ( !JobConstants.isPhantom((short) job)
